requirements.txt
accelerate==1.5.2
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
asttokens==2.4.1
attrs==25.3.0
beautifulsoup4==4.12.3
bertopic==0.17.0
blis==1.3.0
cachetools==5.4.0
catalogue==2.0.10
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.21.0
comm==0.2.2
confection==0.1.5
contourpy==1.3.0
cycler==0.12.1
cymem==2.0.11
db-dtypes==1.3.0
debugpy==1.8.5
decorator==5.1.1
diskcache==5.6.3
executing==2.0.1
fastjsonschema==2.21.1
fastopic==1.0.0
filelock==3.18.0
fonttools==4.53.1
fsspec==2025.3.0
funcy==2.0
gensim==4.3.3
google==3.0.0
google-api-core==2.19.1
google-auth==2.33.0
google-cloud==0.34.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-crc32c==1.5.0
google-genai==1.7.0
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
grpcio==1.65.5
grpcio-status==1.65.5
h11==0.14.0
hdbscan==0.8.40
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.7
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kaleido==0.2.1
kiwisolver==1.4.7
langcodes==3.5.0
language_data==1.3.0
llama_cpp_python==0.3.8
llvmlite==0.44.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.12
narwhals==1.30.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numba==0.61.0
numexpr==2.10.1
numpy==1.26.4
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
plotly==6.0.0
preshed==3.0.9
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==5.27.3
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.11.2
pydantic_core==2.33.1
Pygments==2.18.0
pyLDAvis==3.4.1
pynndescent==0.5.13
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.1.0
referencing==0.36.2
regex==2024.7.24
requests==2.32.3
rich==14.0.0
rpds-py==0.23.1
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
sentence-transformers==3.4.1
setuptools==74.1.2
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
soupsieve==2.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stack-data==0.6.3
sympy==1.13.1
thinc==8.3.6
threadpoolctl==3.5.0
tiktoken==0.9.0
tokenizers==0.21.1
topmost==1.0.2
torch==2.6.0
torchvision==0.21.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.50.0
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2024.1
umap-learn==0.5.7
urllib3==2.2.2
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
websockets==15.0.1
wordcloud==1.9.3
wrapt==1.16.0
